Revert legacy console rename.
